refactor: moved address derive logic to a custom hook
